Refactor generation of predefined symbol sets
Reason: don't repeat yourself. The predefined symbol sets were listed in
two locations, with 2 names (one user-facing, one internal). This commit
rectifies the duplication issue.

We still generate symbol sets programmatically instead of using static
arrays, because doing it that way was a big part of the reason why I
returned to working with this old program in 2017.